# swagger.yaml
openapi: 3.0.3
info:
  title: WhatsApp Multi-Staff Gateway
  description: |
    Self-hosted WhatsApp gateway supporting multiple staff accounts.

    ## Authentication
    When `API_KEY` is configured on the server, pass it via one of:
    - Header: `x-api-key: your-key`
    - Header: `Authorization: Bearer your-key`
    - Query: `?api_key=your-key`

    `/api/health` is always public (no auth required).
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3030
    description: Local development
security:
  - ApiKeyHeader: []
  - BearerAuth: []
  - ApiKeyQuery: []
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
...
paths:
...
  /api/send-text:
    post:
      summary: Send text immediately
      description: |
        Send a text message immediately. WA client for `id` must be connected.
        Rate limited: 30 requests/minute.
      tags: [Immediate Send]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, phone, message]
              properties:
                id:
                  type: string
                  description: Staff ID
                  example: staff1
                phone:
                  type: string
                  description: Phone number (will be normalized — leading 0 becomes 62)
                  example: "08123456789"
                message:
                  type: string
                  example: "Hello from WhatsApp Gateway!"
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  messageId:
                    type: string
                  dbId:
                    type: integer
        '400':
          description: Client not connected or number not registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/send-media:
    post:
      summary: Send media immediately
      description: |
        Send a file or image immediately. `fileData` must be base64-encoded content (without data URI prefix).
        Rate limited: 30 requests/minute.
      tags: [Immediate Send]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, phone, filename, fileData]
              properties:
                id:
                  type: string
                  description: Staff ID
                phone:
                  type: string
                filename:
                  type: string
                  example: "document.pdf"
                fileData:
                  type: string
                  description: Base64-encoded file content
                caption:
                  type: string
                  description: Optional caption
      responses:
        '200':
          description: Media sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  messageId:
                    type: string
                  dbId:
                    type: integer
        '400':
          $ref: '#/components/schemas/Error'
  /api/broadcast-text:
    post:
      summary: Broadcast text to multiple numbers
      description: |
        Send the same text to multiple phone numbers sequentially with a random delay between sends.
        Rate limited: 30 requests/minute.
      tags: [Immediate Send]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, message, phones]
              properties:
                id:
                  type: string
                  description: Staff ID
                message:
                  type: string
                phones:
                  type: array
                  items:
                    type: string
                  example: ["08123456789", "08987654321"]
      responses:
        '200':
          description: Broadcast results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        phone:
                          type: string
                        success:
                          type: boolean
                        id:
                          type: string
                        dbId:
                          type: integer
                        error:
                          type: string
...